+++ extensions/positron-r/amalthea/crates/ark/src/environment/r_environment.rs
@@ -1,4 +1,3 @@
-use std::cmp::Ordering;
 //
 // r_environment.rs
 //
@@ -12,12 +11,12 @@ use crossbeam::channel::select;
 use crossbeam::channel::unbounded;
 use crossbeam::channel::Receiver;
 use crossbeam::channel::Sender;
+use harp::environment::env_bindings;
 use harp::object::RObject;
 use harp::r_lock;
 use harp::r_symbol;
-use harp::symbol::RSymbol;
+use harp::environment::Binding;
 use harp::utils::r_assert_type;
-use harp::utils::r_typeof;
 use libR_sys::*;
 use log::debug;
 use log::error;
@@ -29,53 +28,6 @@ use crate::environment::message::EnvironmentMessageList;
 use crate::environment::message::EnvironmentMessageUpdate;
 use crate::environment::variable::EnvironmentVariable;
 use crate::lsp::signals::SIGNALS;
-
-#[derive(Debug, Eq, PartialEq)]
-enum BindingKind {
-    Regular,
-    // Active,
-    Promise(bool),
-    // Immediate,
-}
-
-#[derive(Debug, Eq, PartialEq)]
-struct Binding {
-    name: RSymbol,
-    binding: SEXP,
-    kind: BindingKind
-}
-
-impl Ord for Binding {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.name.cmp(&other.name)
-    }
-}
-
-impl PartialOrd for Binding {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Binding {
-    pub fn new(frame: SEXP) -> Self {
-
-        let name = RSymbol::new(unsafe {TAG(frame)});
-        let binding = unsafe {CAR(frame)};
-        let kind = match unsafe {r_typeof(binding)} {
-            PROMSXP => BindingKind::Promise(unsafe {PRVALUE(binding) != R_UnboundValue}),
-
-            // TODO: Active and Immediate
-            _ => BindingKind::Regular
-        };
-
-        Self {
-            name,
-            binding,
-            kind
-        }
-    }
-}
 
 /**
  * The R Environment handler provides the server side of Positron's Environment
@@ -355,33 +307,9 @@ impl REnvironment {
     }
 
     fn bindings(&self) -> Vec<Binding> {
-        unsafe {
-            let mut bindings : Vec<Binding> = vec![];
-
-            // 1: traverse the envinronment
-            let env = *self.env;
-            let hash  = HASHTAB(env);
-            if hash == R_NilValue {
-                Self::frame_bindings(FRAME(env), &mut bindings);
-            } else {
-                let n = XLENGTH(hash);
-                for i in 0..n {
-                    Self::frame_bindings(VECTOR_ELT(hash, i), &mut bindings);
-                }
-            }
-
-            // 2: sort by .name i.e. by pointer, not alphabetical order
-            bindings.sort();
-            bindings
-        }
-    }
-
-    unsafe fn frame_bindings(mut frame: SEXP, bindings: &mut Vec<Binding> ) {
-        while frame != R_NilValue {
-            bindings.push(Binding::new(frame));
-
-            frame = CDR(frame);
-        }
+        let mut bindings = env_bindings(self.env.sexp);
+        bindings.sort();
+        bindings
     }
 
 }
