+++ extensions/positron-r/amalthea/crates/harp/src/lib.rs
@@ -18,6 +18,7 @@ pub mod test;
 pub mod utils;
 pub mod vector;
 pub mod symbol;
+pub mod environment;
 
 pub use harp_macros::register;
